generator/app/requirements.txt
azure-cosmos>=4.8.0
azure-identity>=1.15.0
